Moved Action validation to ActionValidator
